Reorganized for client-side SVG rendering
Decided to render the main page content in the browser so it will be
easier to add interactive features in the future: if the browser
already knows where the SVGs are and how they were rendered, it can
more easily modify them or interact with them. Besides, this leaves the
back end functioning mainly as a simple API, which is elegant.

* Installed webpack to handle increased amount of client-side script
* Migrated all content rendering code into client-side scripts,
  rewriting as necessary to create DOM elements instead of raw HTML