Added structural support for special episodes
Instead of merely giving an episode count per season, we give an
"episode map" where "." represents an ordinary episode, "S" represents
a special, and "+" can still be used as a separator as before.

Example: "......+S" is a 6-episode season followed by a special, set
off in its own segment from the main season.

In the data file, seasons are concatenated with "|" into one giant
string mapping the episodes for the entire series.

Actual drawing support for specials is not done yet.